fix: sql i64 error (#891)
<img width="1356" height="280" alt="image"
src="https://github.com/user-attachments/assets/9bb7386a-8039-4684-a4c2-70240a18d7ae"
/>

+++ nt-be/examples/mark_testing_daos.rs
@@ -41,7 +41,7 @@ async fn mark_explicit_daos(
             continue;
         }
 
-        let exists = sqlx::query_scalar::<_, i64>(
+        let exists = sqlx::query_scalar::<_, i32>(
             r#"
             SELECT 1
             FROM monitored_accounts
